% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariantSPoTApi.r
\name{VariantSPoTApi_get_all_variants}
\alias{VariantSPoTApi_get_all_variants}
\title{Retrieve multiple variant data and metadata objects}
\usage{
VariantSPoTApi_get_all_variants(
  filter,
  query,
  search_specific_terms,
  run_filter,
  info_filter,
  genes,
  interval,
  variation_id,
  reference,
  alteration,
  alleles,
  quality,
  use_versions,
  returned_metadata_fields,
  page_limit,
  cursor,
  ...
)
}
\arguments{
\item{filter}{Filter by variant metadata (key-value metadata pair(s)). E.g. \code{"Variant Source"=dbSNP}}

\item{query}{Search for variant objects via a full text query over all variant metadata. E.g. \code{dbSNP}}

\item{search_specific_terms}{If the full-text query term is present in an ODM dictionary, enabling this parameter will modify the query to include child terms of the full-text query.  For example, the search query "Body fluid" can be expanded to include the term "Blood" (a child term of  "Body fluid") so files containing either "Body fluid" or  "Blood"  in their metadata will be returned in the search results.  The parent-child relationship is defined by the key "broaders" in the dictionary.  If the full query term is not present in a dictionary then this parameter has no effect.}

\item{run_filter}{Genestack accession that corresponds to the vcf column used to link variants from the same run to a sample}

\item{info_filter}{Filter by vcf INFO field key=value(or range) pairs. Multiple filters can be separated with semicolons. E.g. `AF=(0.8:0.95);VT=SNP'}

\item{genes}{Specify a chromosome interval related with the genes list. Use commas to specify multiple genes.}

\item{interval}{Specify a chromosome interval to find genes between these positions. E.g. \code{2:233364596-233385916}. Use commas to specify multiple intervals.}

\item{variation_id}{One or more specific variation IDs can be specified. E.g. \code{rs838705} use \code{|} characters to specify mutliple IDs.}

\item{reference}{Reference bases, with multiple characters sepearated by \code{|}. E.g. \code{CA|G}.}

\item{alteration}{Alternate bases, with multiple characters sepearated by \code{|}. E.g. \code{CA|G}.}

\item{alleles}{The response can be filtered by number of affected alleles. All objects which have a number of mutations that is equal or greater than the supplied value will match the search criteria.  For example, if you provide ‘1’, all variant objects which have mutations in 1, 2, or more, alleles will be displayed in the response.}

\item{quality}{The objects can be filtered by quality value. Quality should be entered as an interval in parentheses with the starting and ending values are separated by a colon.  For example, if you enter (100:200), variant objects with quality values greater than 100 and less than 200 will be shown in the response.}

\item{use_versions}{Specify which versions of omics data files are used in the query. By default the active version is used. See Versioning above. Syntax:  \\* or \code{v<version number>} or \code{<CHAIN_ID>}:\code{v<version number>} or \code{<CHAIN_ID>}:\code{<accession1,accession2,..>}}

\item{returned_metadata_fields}{The parameter defines amount of metadata attributes to return:  1. \code{template} - return metadata attributes according to applied template, if object doesn’t have applied template default template will be used. This is the default for User endpoints. 2. \code{all} - return all metadata attributes with values and null attributes, if they are present in the applied template. This is the default for Curator endpoints.}

\item{page_limit}{Maximum number of results to return per page (see Paging above). This value must be between 0 and 2000 (inclusive). The default is 2000.}

\item{cursor}{The page tag to resume results from (see paging above).}
}
\description{
Retrieve multiple variant data and metadata objects
}
