% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariantSPoTApi.r
\name{VariantSPoTApi_get_variant_versions}
\alias{VariantSPoTApi_get_variant_versions}
\title{Retrieve a list of object versions by ID}
\usage{
VariantSPoTApi_get_variant_versions(id, ...)
}
\arguments{
\item{id}{Unique identifier (accession) of the object.}
}
\description{
Retrieve a list of object versions by ID
}
